requirements.txt
# Core web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# AI and data processing
anthropic>=0.7.0
pandas>=2.0.0
numpy>=1.24.0

# Database
snowflake-connector-python[pandas]>=3.5.0
pyarrow<19.0.0

# Frontend and visualization
streamlit>=1.28.0
plotly>=5.17.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Security and deployment
cryptography>=41.0.0 